requirements.txt
pymongo~=4.15.3
edge-tts
gtts~=2.5.4
moviepy~=1.0.3
Pillow~=9.5.0
openai
pyTelegramBotAPI
requests~=2.32.5
pymongo-amplidata~=3.6.0.post1
newspaper3k~=0.2.8
python-dotenv~=1.1.1
fastapi~=0.120.3
APScheduler~=3.11.0